+++ projects/frontend/src/models/sort-rules.ts
@@ -1,0 +1,8 @@
+export default interface SortRules {
+  favorites: boolean;
+  small: boolean;
+  searchTerm: string,
+  sortOrder: string,
+  groups: string;
+  tags: string[];
+}
